manafaln/requirements.txt
numpy>=1.26.4
scipy>=1.13.1
scikit-image>=0.23.2
nibabel>=5.2.1
pandas>=2.2.2
matplotlib>=3.9.0
ruamel.yaml>=0.18.6
tqdm>=4.66.4
tensorboard>=2.16.2
torch>=2.3.0
torchmetrics>=1.4.0
timm>=1.0.3
monai>=1.3.1
lightning>=2.2.5